[b-3] Get all genres Fixes after review.
