+++ pom.xml
@@ -56,6 +56,13 @@
             <scope>provided</scope>
         </dependency>
 
+        <!--Utils-->
+        <dependency>
+            <groupId>javax.annotation</groupId>
+            <artifactId>javax.annotation-api</artifactId>
+            <version>1.3.2</version>
+        </dependency>
+
         <!--Spring-->
         <dependency>
             <groupId>org.springframework</groupId>
