+++ src/main/java/com/filk/service/GenreService.java
@@ -1,11 +1,9 @@
 package com.filk.service;
 
-import com.filk.dao.GenreDao;
 import com.filk.entity.Genre;
 
 import java.util.List;
 
 public interface GenreService {
-    void setGenreDao(GenreDao genreDao);
     List<Genre> getAll();
 }
